Update repository URL in documentation
Ownership of the repository was transferred from GitHub user YuriiSalimov to suoapvs. The URLs in the project
documentation are hereby updated accordingly.

GitHub automatically sets up a redirect when a repository is transferred, so the previous URLs were still functional,
but it is best to avoid the reliance on redirects.